Assignment 7 bo sung chuc nang tim sdt bang username

+++ src/JV2_Ass_7/Main.java
@@ -14,27 +14,46 @@ public class Main{
         phoneDAO phoneDAO= new phoneDAO();
 
         //1. Chuc nang lay ra danh sach
-        ArrayList<PhoneBook> list = phoneDAO.getList();
-        for (int i=0; i<list.size();i++){
-                System.out.println(list.get(i).getId() +"__"+list.get(i).getUser_id()
-                +"__"+list.get(i).getTelephone()+"__"+list.get(i).getStatus());
-            }
+        ArrayList<PhoneBook> listPhone = phoneDAO.getList();
+        for (int i=0; i<listPhone.size();i++){
+            System.out.println(listPhone.get(i).getId() +"__"+listPhone.get(i).getUser_id()
+                    +"__"+listPhone.get(i).getTelephone()+"__"+listPhone.get(i).getStatus());
+        }
 
         //2. them vao danh sach
-        phoneDAO.save(new PhoneBook(list.size()+1,7,"0983151989",ACTIVE));
+        phoneDAO.save(new PhoneBook(listPhone.size()+1,21,"0983151989",ACTIVE));
 
         //3. update
         phoneDAO.update(new PhoneBook(7,3,"087766655",ACTIVE));
 
         //4.Delete
-        Integer id = 7;
-        PhoneBook phoneBook_delete= new PhoneBook();
-        phoneBook_delete.setId(id);
-        phoneBook_delete.setUser_id(list.get(id).getUser_id());
-        phoneBook_delete.setTelephone(list.get(id).getTelephone());
-        phoneBook_delete.setStatus(list.get(id).getStatus());
+        Integer id = 37;
+        try {
+            if(listPhone.size()+1>id ){
+                PhoneBook phoneBook_delete= new PhoneBook(
+                        id,
+                        listPhone.get(id).getUser_id(),
+                        listPhone.get(id).getTelephone(),
+                        listPhone.get(id).getStatus()
+                );
+                phoneDAO.delete(phoneBook_delete);
+            }else {
+                System.out.println("Vuot qua do dai database");
+            }
+        }catch (Exception e){e.printStackTrace();};
 
-        phoneDAO.delete(phoneBook_delete);
+        //5. List user
+        UserDAO userDAO= new UserDAO();
+        ArrayList<User> listUser = userDAO.getList();
+        //in ra
+        for(int i=0;i<listUser.size();i++){
+            System.out.println(listUser.get(i).getId() +"__"+listUser.get(i).getUsername());
+        }
+
+        //6.Tim sdt cua user
+        FindPhone findPhone = new FindPhone();
+        findPhone.tim("Hoa");
+
     }
 
 }
